+++ app/src/main/java/com/easyhz/patchnote/ui/screen/defect/export/DefectExportViewModel.kt
@@ -1,4 +1,4 @@
-package com.easyhz.patchnote.ui.screen.export
+package com.easyhz.patchnote.ui.screen.defect.export
 
 import android.content.Context
 import android.os.Build
@@ -16,10 +16,10 @@ import com.easyhz.patchnote.core.model.filter.FilterValue.Companion.asInt
 import com.easyhz.patchnote.core.model.filter.FilterValue.Companion.asLong
 import com.easyhz.patchnote.core.model.filter.FilterValue.Companion.asString
 import com.easyhz.patchnote.domain.usecase.defect.ExportDefectUseCase
-import com.easyhz.patchnote.ui.screen.export.contract.DefectExportIntent
-import com.easyhz.patchnote.ui.screen.export.contract.DefectExportSideEffect
-import com.easyhz.patchnote.ui.screen.export.contract.DefectExportState
-import com.easyhz.patchnote.ui.screen.export.contract.DefectExportState.Companion.updateFilterItemValue
+import com.easyhz.patchnote.ui.screen.defect.export.contract.DefectExportIntent
+import com.easyhz.patchnote.ui.screen.defect.export.contract.DefectExportSideEffect
+import com.easyhz.patchnote.ui.screen.defect.export.contract.DefectExportState
+import com.easyhz.patchnote.ui.screen.defect.export.contract.DefectExportState.Companion.updateFilterItemValue
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.Dispatchers
