+++ app/src/main/java/com/easyhz/patchnote/ui/screen/defect/export/contract/DefectExportIntent.kt
@@ -1,4 +1,4 @@
-package com.easyhz.patchnote.ui.screen.export.contract
+package com.easyhz.patchnote.ui.screen.defect.export.contract
 
 import com.easyhz.patchnote.core.common.base.UiIntent
 import com.easyhz.patchnote.core.model.filter.Filter
