+++ app/src/main/java/com/easyhz/patchnote/ui/screen/defect/offline/detail/contract/OfflineDefectDetailIntent.kt
@@ -1,4 +1,4 @@
-package com.easyhz.patchnote.ui.screen.offline.detail.contract
+package com.easyhz.patchnote.ui.screen.defect.offline.detail.contract
 
 import com.easyhz.patchnote.core.common.base.UiIntent
 import com.easyhz.patchnote.core.designSystem.util.bottomSheet.DefectDetailBottomSheet
