+++ app/src/main/java/com/easyhz/patchnote/ui/screen/filter/FilterScreen.kt
@@ -48,7 +48,7 @@ import com.easyhz.patchnote.core.model.filter.FilterValue.Companion.asString
 import com.easyhz.patchnote.ui.screen.defect.common.DefectViewModel
 import com.easyhz.patchnote.ui.screen.defect.common.contract.DefectIntent
 import com.easyhz.patchnote.ui.screen.defect.common.contract.DefectSideEffect
-import com.easyhz.patchnote.ui.screen.defectEntry.component.DefectCategoryField
+import com.easyhz.patchnote.ui.screen.defect.defectEntry.component.DefectCategoryField
 import com.easyhz.patchnote.ui.screen.filter.contract.FilterIntent
 import com.easyhz.patchnote.ui.screen.filter.contract.FilterSideEffect
 import com.easyhz.patchnote.ui.theme.MainText
